services/vanna/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pydantic==2.9.2
vanna==0.7.5
groq==0.11.0
